feat: Move all API related files into functions/api.
